+++ views/main.py
@@ -40,7 +40,7 @@ from internship.decorators.cohort_view_decorators import redirect_if_not_in_coho
 from internship.decorators.global_view_decorators import redirect_if_multiple_registrations
 from internship.decorators.score_encoding_view_decorators import redirect_if_not_master
 from internship.models.enums.user_account_status import UserAccountStatus
-from internship.views.api_client import get_master_by_email, activate_master_account
+from internship.services.internship import InternshipAPIService
 
 
 @login_required
@@ -50,11 +50,11 @@ def view_internship_role_selection(request):
     except MultipleObjectsReturned:
         return dash_main_view.show_multiple_registration_id_error(request)
 
-    master = get_master_by_email(email=request.user.email)
+    master = InternshipAPIService.get_master_by_email(email=request.user.email)
 
     if master:
         if master['user_account_status'] != UserAccountStatus.ACTIVE:
-            activate_master_account(master['uuid'])
+            InternshipAPIService.activate_master_account(master['uuid'])
         return redirect(reverse('internship_master_home'))
     elif student:
         return redirect(reverse('internship_cohort_selection'))
